feat(query): allow returning multiple leaf nodes in a single result
Thus we can have a query where each result is a sentence unit and the
0 or more word units it contains, if the word node has the
configuration option `multiple = true`.

There can only be one node marked with `multiple` and it must not have
any children in the query.

Additionally, this commit adds `WhereClause`, a wrapper class for
generating queries that abstracts over the difference between single
items and lists (and may do other things in future).